Bump metadata (#4558)

+++ packages/api-augment/src/kusama/consts.ts
@@ -592,8 +592,13 @@ declare module '@polkadot/api-base/types/consts' {
       friendDepositFactor: u128 & AugmentedConst<ApiType>;
       /**
        * The maximum amount of friends allowed in a recovery configuration.
+       * 
+       * NOTE: The threshold programmed in this Pallet uses u16, so it does
+       * not really make sense to have a limit here greater than u16::MAX.
+       * But also, that is a lot more than you should probably set this value
+       * to anyway...
        **/
-      maxFriends: u16 & AugmentedConst<ApiType>;
+      maxFriends: u32 & AugmentedConst<ApiType>;
       /**
        * The base amount of currency needed to reserve for starting a recovery.
        * 
@@ -846,6 +851,8 @@ declare module '@polkadot/api-base/types/consts' {
       burn: Permill & AugmentedConst<ApiType>;
       /**
        * The maximum number of approvals that can wait in the spending queue.
+       * 
+       * NOTE: This parameter is also used within the Bounties Pallet extension if enabled.
        **/
       maxApprovals: u32 & AugmentedConst<ApiType>;
       /**
